move collision checking code from utils to sim_utils
