+++ config/sim_config/actuated_cylinder_sim_config.yaml
@@ -80,7 +80,7 @@ data_collection_config:
   _target_: planning_through_contact.simulation.environments.data_collection_table_environment.DataCollectionConfig
   generate_plans: false
   render_plans: true
-  convert_to_zarr: false
+  convert_to_zarr: true
   convert_to_zarr_reduce: false # Supercloud flag to perform "reduce" step
   plans_dir: 'temp/plans'
   rendered_plans_dir: 'temp/rendered'
