+++ planning_through_contact/experiments/utils.py
@@ -1,14 +1,9 @@
-from typing import Literal, Optional, Tuple, List
+from typing import Literal, Optional, Tuple
 
 import numpy as np
 
 from planning_through_contact.geometry.collision_geometry.box_2d import Box2d
 from planning_through_contact.geometry.collision_geometry.t_pusher_2d import TPusher2d
-from planning_through_contact.geometry.collision_geometry.collision_geometry import (
-    CollisionGeometry,
-    ContactLocation,
-    PolytopeContactLocation,
-)
 from planning_through_contact.geometry.planar.planar_pose import PlanarPose
 from planning_through_contact.geometry.rigid_body import RigidBody
 from planning_through_contact.planning.planar.planar_plan_config import (
@@ -21,10 +16,6 @@ from planning_through_contact.planning.planar.planar_plan_config import (
     PlanarPushingWorkspace,
     PlanarSolverParams,
     SliderPusherSystemConfig,
-)
-
-from planning_through_contact.geometry.planar.non_collision import (
-    check_finger_pose_in_contact_location,
 )
 
 def get_box() -> RigidBody:
@@ -229,110 +220,3 @@ def sample_random_plan(
         slider_initial_pose, slider_target_pose, pusher_pose, pusher_pose
     )
     return plan
-
-# Util functions for getting initial slider poses
-
-def get_slider_start_poses(
-    seed: int,
-    num_plans: int,
-    workspace: PlanarPushingWorkspace,
-    config: PlanarPlanConfig,
-    pusher_pose: PlanarPose,
-    limit_rotations: bool = True,  # Use this to start with
-) -> List[PlanarPushingStartAndGoal]:
-    # We want the plans to always be the same
-    np.random.seed(seed)
-    slider = config.slider_geometry
-    slider_initial_poses = []
-    for _ in range(num_plans):
-        slider_initial_pose = _get_slider_pose_within_workspace(
-            workspace, slider, pusher_pose, config, limit_rotations
-        )
-        slider_initial_poses.append(slider_initial_pose)
-
-    return slider_initial_poses
-
-def _get_slider_pose_within_workspace(
-    workspace: PlanarPushingWorkspace,
-    slider: CollisionGeometry,
-    pusher_pose: PlanarPose,
-    config: PlanarPlanConfig,
-    limit_rotations: bool = False,
-    enforce_entire_slider_within_workspace: bool = True,
-) -> PlanarPose:
-    valid_pose = False
-
-    slider_pose = None
-    while not valid_pose:
-        x_initial = np.random.uniform(workspace.slider.x_min, workspace.slider.x_max)
-        y_initial = np.random.uniform(workspace.slider.y_min, workspace.slider.y_max)
-        EPS = 0.01
-        if limit_rotations:
-            # th_initial = np.random.uniform(-np.pi / 2 + EPS, np.pi / 2 - EPS)
-            th_initial = np.random.uniform(-np.pi / 4 + EPS, np.pi / 4 - EPS)
-        else:
-            th_initial = np.random.uniform(-np.pi + EPS, np.pi - EPS)
-
-        slider_pose = PlanarPose(x_initial, y_initial, th_initial)
-
-        collides_with_pusher = _check_collision(pusher_pose, slider_pose, config)
-        within_workspace = _slider_within_workspace(workspace, slider_pose, slider)
-
-        if enforce_entire_slider_within_workspace:
-            valid_pose = within_workspace and not collides_with_pusher
-        else:
-            valid_pose = not collides_with_pusher
-
-    assert slider_pose is not None  # fix LSP errors
-
-    return slider_pose
-
-# TODO: refactor
-def _check_collision(
-    pusher_pose_world: PlanarPose,
-    slider_pose_world: PlanarPose,
-    config: PlanarPlanConfig,
-) -> bool:
-    p_WP = pusher_pose_world.pos()
-    R_WB = slider_pose_world.two_d_rot_matrix()
-    p_WB = slider_pose_world.pos()
-
-    # We need to compute the pusher pos in the frame of the slider
-    p_BP = R_WB.T @ (p_WP - p_WB)
-    pusher_pose_body = PlanarPose(p_BP[0, 0], p_BP[1, 0], 0)
-
-    # we always add all non-collision modes, even when we don't add all contact modes
-    # (think of maneuvering around the object etc)
-    locations = [
-        PolytopeContactLocation(ContactLocation.FACE, idx)
-        for idx in range(config.slider_geometry.num_collision_free_regions)
-    ]
-    matching_locs = [
-        loc
-        for loc in locations
-        if check_finger_pose_in_contact_location(pusher_pose_body, loc, config)
-    ]
-    if len(matching_locs) == 0:
-        return True
-    else:
-        return False
-    
-def _slider_within_workspace(
-    workspace: PlanarPushingWorkspace, pose: PlanarPose, slider: CollisionGeometry
-) -> bool:
-    """
-    Checks whether the entire slider is within the workspace
-    """
-    R_WB = pose.two_d_rot_matrix()
-    p_WB = pose.pos()
-
-    p_Wv_s = [
-        slider.get_p_Wv_i(vertex_idx, R_WB, p_WB).flatten()
-        for vertex_idx in range(len(slider.vertices))
-    ]
-
-    lb, ub = workspace.slider.bounds
-    vertices_within_workspace: bool = np.all([v <= ub for v in p_Wv_s]) and np.all(
-        [v >= lb for v in p_Wv_s]
-    )
-    return vertices_within_workspace
